+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 - `[General]` Added 4.90 release
 - `[Datagrid]` Added fix to button types and errors in wizard. ([#1598](https://github.com/infor-design/enterprise-ng/issues/1598))
 - `[Datagrid]` Fixed tab key navigation when using actionable mode when having editor or formatter. ([EP#8141](https://github.com/infor-design/enterprise/issues/8141))
+- `[Pie/Donut]` Fixed rendering issues when having bordered class in the widget. ([EP#8164](https://github.com/infor-design/enterprise/issues/8164))
 
 ## 17.0.0